fix error

+++ src/assets/pages/userData/muser/muserList.vue
@@ -111,11 +111,19 @@ export default {
             this.getMUserList();
         },
         handleDelete(id) {
-            const ids = [id];
-            API.deleteMUser(JSON.stringify(ids)).then(res => {
-                this.$message('删除成功!');
-                this.getMUserList();
-            })
+            this.$confirm('确定删除吗?', '提示', {
+                confirmButtonText: '确定',
+                cancelButtonText: '取消',
+                type: 'warning'
+            }).then(() => {
+                const ids = [id];
+                API.deleteMUser(JSON.stringify(ids)).then(res => {
+                    this.$message('删除成功!');
+                    this.getMUserList();
+                })
+            }).catch(() => {
+                
+            });
         }
     },
     watch: {
